reduce saml diff
